requirements.txt
langchain>=0.3
langchain-groq>=0.2
langchain-community>=0.3
langchain-chroma>=0.1
langgraph>=0.2
langsmith>=0.1
langfuse>=2.0
chromadb>=0.5
tavily-python>=0.3
beautifulsoup4>=4.12
requests>=2.31
streamlit>=1.35
python-dotenv>=1.0
sentence-transformers>=3.0
fpdf2>=2.7
